requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
python-dotenv
streamlit
faiss-cpu
beautifulsoup4
